refactor: migrate to Prisma, optimize GitHub sync, and add import by link - Migrated data access layer from raw SQL to Prisma ORM for better type safety and maintainability. - Optimized repository import and refresh logic by implementing bulk fetching and in-memory lookups, solving the N+1 query problem. - Added "Import Repository by Link" feature to allow adding any public GitHub repository via URL. - Updated GitHub sync to automatically move closed issues to 'done' and handle reopened issues. - Refactored the large Kanban component into a modular folder structure. - Fixed UI bugs related to drag-and-drop target areas and event propagation. - Updated Docker and build configuration to automate Prisma client generation and schema syncing.
